provide (parse s), which parses rfc3339-style timestamp
      - time stamp can elide trailing components
      - time-offset is optional
      - an absent time-offset is taken to be UTC (i.e. treated as if
        +00:00 had been specified.)

Parse is pluggable, the sense that it takes a constructor function which
is called with the the result of the parse. The value returned by the
constructor is ultimately the result of the call to parse.

constructors are provided for the Java types java.util.Date,
java.util.Calendar and java.sql.Timestamp.

Functions are provided to format each of these three types
as rfc3339 timestamps, with some possible minor lossage
of information depending on which type is chosen:

- Calendar preserves only milliseconds
- Date preserves only milliseconds and forgets the timezone
  offset the original literal used.
- Timestamp preserves nanoseconds (yea), but forgets the
  timezone.

When I say "forgets the timezone" I mean that if we
format the object in question back to an rfc3339
timestamp, we'll end up with something like:

2000-01-02T12:00:00+00:00 and it won't be possible
to tell if the original literal ended with 13:00:00+01:00
or maybe 06:00:00-06:00. (All three forms specify the
same instant in time, so there's no real lossage.)